refactor: hoist pagination + strict ISO datetime helpers to core; add error codes
