+++ src/synthorg/core/error_taxonomy.py
@@ -107,6 +107,12 @@ class ErrorCode(IntEnum):
     DELIVERABLE_RECEIPT_NOT_FOUND = 3023
     API_KEY_NOT_FOUND = 3024
     UPGRADE_RECOMMENDATION_NOT_FOUND = 3025
+    GROUP_PARTICIPANT_UNKNOWN = 3026
+    CHECKPOINT_NOT_FOUND = 3027
+    FINE_TUNE_RUN_NOT_FOUND = 3028
+    AGENT_NOT_FOUND = 3029
+    PERSONALITY_NOT_FOUND = 3030
+    TRAINING_SESSION_NOT_FOUND = 3031
 
     # 4xxx -- conflict
     RESOURCE_CONFLICT = 4000
@@ -211,6 +217,11 @@ class ErrorCode(IntEnum):
     SINK_CONSTRUCTION_ERROR = 8036
     SUBWORKFLOW_CYCLE_ERROR = 8037
     SUBWORKFLOW_DEPTH_EXCEEDED_ERROR = 8038
+    TASK_ENGINE_NOT_RUNNING = 8039
+    TASK_ENGINE_QUEUE_FULL = 8040
+    CONVERSATIONAL_PROPOSE_UNAVAILABLE = 8041
+    GROUP_CHAT_UNAVAILABLE = 8042
+    CHARTER_INTERVIEW_UNAVAILABLE = 8043
 
 
 # Error-code band for the NOT_FOUND category (3xxx).  Per-domain
